refactor: consolidate google connectors + archive legacy GTK UI
## bantz.google — single package for all Google services

connectors/google/ → google/ (all files moved):
- auth_manager.py      (was connectors/google/auth_manager.py)
- connector_base.py    (was connectors/google/base.py)
- classroom.py         (was connectors/google/classroom.py)
- contacts_connector.py (was connectors/google/contacts.py)
- keep.py              (was connectors/google/keep.py)
- tasks.py             (was connectors/google/tasks.py)
- entity_linker.py     (was connectors/google/entity_linker.py)

connectors/google/__init__.py kept as backward-compat shim
(re-exports GoogleAuthManager + GoogleConnector from bantz.google)

Updated imports in:
- src/bantz/google/*      (internal references)
- src/bantz/data/google_sync.py
- src/bantz/tools/register_all.py
- bantz-overlay/src/main/main.js (Python script strings)
- tests/test_issue_1292_google_suite.py
- tests/test_classroom_tools.py

## bantz.ui — GTK overlay archived

src/bantz/ui/*.py → _legacy/ui/ (25 files)
Desktop HUD is now handled by bantz-overlay/ (Electron).
src/bantz/ui/__init__.py kept as empty stub (prevents ImportError
in TYPE_CHECKING blocks in agent/controller.py).
