+++ _legacy/ui/components/mini_terminal.py
