+++ _legacy/ui/components/status_bar.py
