+++ _legacy/ui/components/waveform.py
