+++ _legacy/ui/image_slot.py
