+++ _legacy/ui/panel_animator.py
