+++ _legacy/ui/streaming/highlighter.py
