+++ _legacy/ui/ticker.py
