+++ src/bantz/google/gmail_auth.py
@@ -174,7 +174,7 @@ def get_gmail_credentials(
     # ── Unified auth manager bridge (Issue #1292) ───────────────
     if client_secret_path is None and token_path is None:
         try:
-            from bantz.connectors.google.auth_manager import get_auth_manager
+            from bantz.google.auth_manager import get_auth_manager
 
             mgr = get_auth_manager()
             if mgr is not None:
