+++ src/bantz/google/keep.py
@@ -12,7 +12,7 @@ import logging
 from dataclasses import dataclass, field
 from typing import Any, Optional
 
-from bantz.connectors.google.base import GoogleConnector, ToolSchema
+from bantz.google.connector_base import GoogleConnector, ToolSchema
 
 logger = logging.getLogger(__name__)
